attempted to use GoalCont for parts of user reg, make goal and user. Not working

+++ Laravel/Trkr/app/Http/Controllers/user/GoalController.php
@@ -1,13 +1,19 @@
 <?php
 
 namespace App\Http\Controllers\User;
+
 use App\Http\Controllers\Controller;
+use App\Models\Country;
 use App\Models\Goal;
+use App\Models\Interest;
+use App\Models\Role;
 use App\Models\Task;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Session;
 use Illuminate\Support\Str;
+use Illuminate\Database\Eloquent\Builder;
 
 class GoalController extends Controller
 {
@@ -21,12 +27,18 @@ class GoalController extends Controller
         $user = Auth::user();   // Checking if the user is logged in
         $user->authorizeRoles('user'); // Checking if the user is an admin
 
+        // Authorise user first
+        if ($goal->user->id != Auth::id()) {
+            //403 error forbidden
+            return abort(403);
+        }
+
         // Fetch goals in order of when they were created and limited to 5 per page
         $goals = Goal::where('user_id', Auth::id())
-        ->latest('updated_at')
-        ->with('tasks')
-        ->with('user')
-        ->paginate(5);
+            ->latest('updated_at')
+            ->with('tasks')
+            ->with('user')
+            ->paginate(5);
 
         // Returns the goals index view and passes the goals variable with the logged in users' goals
         return view('user.goals.index')->with('goals', $goals);
@@ -39,14 +51,26 @@ class GoalController extends Controller
      */
     public function create()
     {
-        $user = Auth::user();  
+        $user = Auth::user();
         $user->authorizeRoles('user');
 
-         // Fetch goals in order of when they were created and limited to 5 per page
-         $goals = Goal::where('user_id', Auth::id())->get();
-         $languages = ['German', 'Spanish', 'French', 'Italian'];
+        // Gets all users that have role 'user', no admins returned
+        $partners = User::whereNot('id', $user->id)
+            ->with('roles')
+            ->whereHas('roles', function (Builder $query) {
+                $query->where('name', 'user');
+            })
+            ->limit(8)
+            ->get();
 
-        return view ('user.goals.create')->with('goals', $goals)->with('languages', $languages);
+        // Fetch goals in order of when they were created and limited to 5 per page
+        $goals = Goal::where('user_id', Auth::id())->get();
+        $languages = ['German', 'Spanish', 'French', 'Italian'];
+        $countries = Country::all();
+        $interests = Interest::all();
+
+        return view('user.goals.create')->with(compact('goals', 'languages', 'partners', 'countries', 'interests', 'user'));
+        // return view('user.goals.create')->with('goals', $goals)->with('languages', $languages);
     }
 
     /**
@@ -55,16 +79,23 @@ class GoalController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(Request $request, User $user)
     {
-        $user = Auth::user();  
-        $user->authorizeRoles('user');
+        $user = User::where('id', (int)$request->user)->get();
+        // dd($user);
+
+        $userAuth = Auth::user();
+        $userAuth->authorizeRoles('user');
+
+        // dd($request);
 
         $request->validate([
             'title' => 'required|max:50',
             'description' => 'required',
             // Issue with validating select option, 
-            'language' => 'required'
+            'about_me' => 'required',
+            'country_id' => 'required',
+            'language' => 'required',
         ]);
 
         Goal::create([
@@ -74,7 +105,7 @@ class GoalController extends Controller
             'language' => $request->language
         ]);
 
-        return to_route('user.goals.index');
+        return to_route('home.dashboard')->with('toast_success', 'Goal Created Successfully!');
     }
 
     /**
@@ -85,17 +116,23 @@ class GoalController extends Controller
      */
     public function show(Goal $goal)
     {
-        $user = Auth::user();  
+        $user = Auth::user();
         $user->authorizeRoles('user');
+
+        // Authorise user first
+        if ($goal->user->id != Auth::id()) {
+            //403 error forbidden
+            return abort(403);
+        }
 
         $toDo = Task::where('status', 0)->where('goal_id', $goal->id)->get();
         $done = Task::where('status', 1)->where('goal_id', $goal->id)->get();
         $user = Auth::user();   // Not needed. test removing
 
-         // Put url into session data to redirect back after editing task
-         Session::put('url', request()->fullUrl());
+        // Put url into session data to redirect back after editing task
+        Session::put('url', request()->fullUrl());
 
-        return view('user.goals.show', with(["goal" =>$goal, "toDo" => $toDo, "done"=> $done, "user" => $user]));
+        return view('user.goals.show', with(["goal" => $goal, "toDo" => $toDo, "done" => $done, "user" => $user]));
     }
 
     /**
@@ -106,8 +143,15 @@ class GoalController extends Controller
      */
     public function edit(Goal $goal)
     {
-        $user = Auth::user();  
-        $user->authorizeRoles('user');
+
+        $userAuth = Auth::user();
+        $userAuth->authorizeRoles('user');
+
+        // Authorise user first
+        if ($goal->user->id != Auth::id()) {
+            //403 error forbidden
+            return abort(403);
+        }
 
         // TODO Create a languages table, easier to manage and update
         $languages = ['German', 'Spanish', 'French', 'Italian'];
@@ -124,8 +168,15 @@ class GoalController extends Controller
      */
     public function update(Request $request, Goal $goal)
     {
-        $user = Auth::user();  
-        $user->authorizeRoles('user');
+        // dd("update");
+        $userAuth = Auth::user();
+        $userAuth->authorizeRoles('user');
+
+        // Authorise user first
+        if ($goal->user->id != Auth::id()) {
+            //403 error forbidden
+            return abort(403);
+        }
 
         // TODO: Add validation on 'language' to confirm it's a valid option
         $request->validate([
@@ -140,6 +191,13 @@ class GoalController extends Controller
             'language' => $request->language
         ]);
 
+        // If the user comes from Dashboard, session data will exist for it, so redirect there
+        if (session('url')) {
+            $url = session('url');
+            $request->session()->forget('url');
+            return redirect($url);
+        }
+
         return to_route('user.goals.show', $goal)->with('toast_success', 'Goal Updated Successfully!');
     }
 
@@ -151,12 +209,18 @@ class GoalController extends Controller
      */
     public function destroy(Goal $goal)
     {
-        $user = Auth::user();  
-        $user->authorizeRoles('user');
+        // dd("delete");
+        $userAuth = Auth::user();
+        $userAuth->authorizeRoles('user');
+
+        // Authorise user first
+        if ($goal->user->id != Auth::id()) {
+            //403 error forbidden
+            return abort(403);
+        }
 
         $goal->delete();
 
         return to_route('user.goals.index');
     }
 }
-
